S2EExecutor: moved S2EExternalDispatcher class to separate file
