+++ src/S2EExecutor.cpp
@@ -10,6 +10,7 @@
 #include <s2e/S2E.h>
 #include <s2e/S2EExecutionState.h>
 #include <s2e/S2EExecutor.h>
+#include <s2e/S2EExternalDispatcher.h>
 #include <s2e/SymbolicHardwareHook.h>
 #include <s2e/Utils.h>
 #include <s2e/cpu.h>
@@ -20,8 +21,6 @@
 
 #include <s2e/s2e_libcpu.h>
 
-#include <llvm/ExecutionEngine/ExecutionEngine.h>
-#include <llvm/ExecutionEngine/GenericValue.h>
 #include <llvm/IR/Constants.h>
 #include <llvm/IR/DataLayout.h>
 #include <llvm/IR/DerivedTypes.h>
@@ -238,111 +237,6 @@ namespace s2e {
 
 /* Global array to hold tb function arguments */
 volatile void *tb_function_args[3];
-
-/* External dispatcher to convert longjmp's into C++ exceptions */
-class S2EExternalDispatcher : public klee::ExternalDispatcher {
-protected:
-    virtual bool runProtectedCall(llvm::Function *f, uint64_t *args);
-
-public:
-    S2EExternalDispatcher(llvm::LLVMContext &context) : ExternalDispatcher(context) {
-    }
-
-    void removeFunction(llvm::Function *f);
-};
-
-extern "C" {
-
-// FIXME: This is not reentrant.
-static jmp_buf s2e_escapeCallJmpBuf;
-static jmp_buf s2e_cpuExitJmpBuf;
-
-#ifdef _WIN32
-static void s2e_ext_sigsegv_handler(int signal) {
-}
-#else
-static void s2e_ext_sigsegv_handler(int signal, siginfo_t *info, void *context) {
-    longjmp(s2e_escapeCallJmpBuf, 1);
-}
-#endif
-}
-
-bool S2EExternalDispatcher::runProtectedCall(Function *f, uint64_t *args) {
-#ifndef _WIN32
-    struct sigaction segvAction, segvActionOld;
-#endif
-    bool res;
-
-    if (!f)
-        return false;
-
-    gTheArgsP = args;
-
-#ifdef _WIN32
-    signal(SIGSEGV, s2e_ext_sigsegv_handler);
-#else
-    segvAction.sa_handler = 0;
-    memset(&segvAction.sa_mask, 0, sizeof(segvAction.sa_mask));
-    segvAction.sa_flags = SA_SIGINFO;
-    segvAction.sa_sigaction = s2e_ext_sigsegv_handler;
-    sigaction(SIGSEGV, &segvAction, &segvActionOld);
-#endif
-
-    memcpy(s2e_cpuExitJmpBuf, env->jmp_env, sizeof(env->jmp_env));
-
-    if (setjmp(env->jmp_env)) {
-        memcpy(env->jmp_env, s2e_cpuExitJmpBuf, sizeof(env->jmp_env));
-        throw CpuExitException();
-    } else {
-        if (setjmp(s2e_escapeCallJmpBuf)) {
-            res = false;
-        } else {
-            std::vector<GenericValue> gvArgs;
-            ExecutionEngine *ee = getExecutionEngine(f);
-
-            ee->runFunction(f, gvArgs);
-            res = true;
-        }
-    }
-
-    memcpy(env->jmp_env, s2e_cpuExitJmpBuf, sizeof(env->jmp_env));
-
-#ifdef _WIN32
-#warning Implement more robust signal handling on windows
-    signal(SIGSEGV, SIG_IGN);
-#else
-    sigaction(SIGSEGV, &segvActionOld, 0);
-#endif
-    return res;
-}
-
-/**
- * Remove all mappings between calls to external functions
- * and the actual external call stub generated by KLEE.
- * Also remove the machine code for the stub and the stub itself.
- * This is used whenever S2E deletes a translation block and its LLVM
- * representation. Failing to do so would leave stale references to
- * machine code in KLEE's external dispatcher.
- */
-void S2EExternalDispatcher::removeFunction(llvm::Function *f) {
-    dispatchers_ty::iterator it, itn;
-
-    it = dispatchers.begin();
-    while (it != dispatchers.end()) {
-        if ((*it).first->getParent()->getParent() == f) {
-
-            llvm::Function *dispatcher = (*it).second;
-            dispatcher->eraseFromParent();
-
-            itn = it;
-            ++itn;
-            dispatchers.erase(it);
-            it = itn;
-        } else {
-            ++it;
-        }
-    }
-}
 
 S2EExecutor::S2EExecutor(S2E *s2e, TCGLLVMContext *tcgLLVMContext, InterpreterHandler *ie)
     : Executor(ie, tcgLLVMContext->getLLVMContext()), m_s2e(s2e), m_tcgLLVMContext(tcgLLVMContext),
@@ -1309,16 +1203,16 @@ uintptr_t S2EExecutor::executeTranslationBlockConcrete(S2EExecutionState *state,
     ++state->m_stats.m_statTranslationBlockConcrete;
 
     uintptr_t ret = 0;
-    memcpy(s2e_cpuExitJmpBuf, env->jmp_env, sizeof(env->jmp_env));
+    S2EExternalDispatcher::saveJmpBuf();
 
     if (setjmp(env->jmp_env)) {
-        memcpy(env->jmp_env, s2e_cpuExitJmpBuf, sizeof(env->jmp_env));
+        S2EExternalDispatcher::restoreJmpBuf();
         throw CpuExitException();
     } else {
         ret = tcg_libcpu_tb_exec(env, tb->tc_ptr);
     }
 
-    memcpy(env->jmp_env, s2e_cpuExitJmpBuf, sizeof(env->jmp_env));
+    S2EExternalDispatcher::restoreJmpBuf();
     return ret;
 }
